Refactor ChatPurge
+ Add Try/Catches
+ Add More Logging
+ Add ContentType Checking
* Move to privacy Folder
