+++ app/src/main/kotlin/me/rhunk/snapenhance/manager/impl/FeatureManager.kt
@@ -24,7 +24,7 @@ import me.rhunk.snapenhance.features.impl.spying.StealthMode
 import me.rhunk.snapenhance.features.impl.tweaks.AntiAutoSave
 import me.rhunk.snapenhance.features.impl.tweaks.AutoSave
 import me.rhunk.snapenhance.features.impl.tweaks.CameraTweaks
-import me.rhunk.snapenhance.features.impl.tweaks.ChatPurge
+import me.rhunk.snapenhance.features.impl.privacy.ChatPurge
 import me.rhunk.snapenhance.features.impl.tweaks.DisableVideoLengthRestriction
 import me.rhunk.snapenhance.features.impl.tweaks.GalleryMediaSendOverride
 import me.rhunk.snapenhance.features.impl.tweaks.GooglePlayServicesDialogs
